# Accelerate ROIPooling layer