--- a/src/operator/roi_pooling.cu
+++ b/src/operator/roi_pooling.cu
@@ -34,8 +34,8 @@ namespace cuda {
 
 template<typename Dtype>
 __global__ void ROIPoolForwardKernel(const int count, const Dtype* bottom_data,
-                                     const float spatial_scale, const int channels,
-                                     const int height, const int width,
+                                     const float spatial_scale, const int batch_size,
+                                     const int channels, const int height, const int width,
                                      const int pooled_height, const int pooled_width,
                                      const Dtype* bottom_rois, Dtype* top_data,
                                      Dtype* argmax_data) {
@@ -51,9 +51,9 @@ __global__ void ROIPoolForwardKernel(const int count, const Dtype* bottom_data,
     bottom_rois += n * 5;
     int roi_batch_ind = bottom_rois[0];
 
-    if (roi_batch_ind < 0) {
+    if (roi_batch_ind < 0 || roi_batch_ind >= batch_size) {
       top_data[index] = 0;
-      argmax_data[index] = 0;
+      argmax_data[index] = -1;
       continue;
     }
 
@@ -90,13 +90,14 @@ __global__ void ROIPoolForwardKernel(const int count, const Dtype* bottom_data,
     Dtype maxval = is_empty ? 0 : -FLT_MAX;
     // If nothing is pooled, argmax = -1 causes nothing to be backprop'd
     int maxidx = -1;
-    bottom_data += (roi_batch_ind * channels + c) * height * width;
+    int offset_bottom_data = (roi_batch_ind * channels + c) * height * width;
+    bottom_data += offset_bottom_data;
     for (int h = hstart; h < hend; ++h) {
       for (int w = wstart; w < wend; ++w) {
         int bottom_index = h * width + w;
         if (bottom_data[bottom_index] > maxval) {
           maxval = bottom_data[bottom_index];
-          maxidx = bottom_index;
+          maxidx = offset_bottom_data + bottom_index;
         }
       }
     }
@@ -116,6 +117,7 @@ inline void ROIPoolForward(const Tensor<gpu, 4, Dtype> &out,
   Dtype *top_data = out.dptr_;
   Dtype *argmax_data = max_idx.dptr_;
   const int count = out.shape_.Size();
+  const int batch_size = data.size(0);
   const int channels = data.size(1);
   const int height = data.size(2);
   const int width = data.size(3);
@@ -127,7 +129,7 @@ inline void ROIPoolForward(const Tensor<gpu, 4, Dtype> &out,
   CheckLaunchParam(dimGrid, dimBlock, "ROIPooling Forward");
   cudaStream_t stream = Stream<gpu>::GetStream(out.stream_);
   ROIPoolForwardKernel<Dtype><<<dimGrid, dimBlock, 0, stream>>>(
-      count, bottom_data, spatial_scale, channels, height, width,
+      count, bottom_data, spatial_scale, batch_size, channels, height, width,
       pooled_height, pooled_width, bottom_rois, top_data, argmax_data);
   MSHADOW_CUDA_POST_KERNEL_CHECK(ROIPoolForwardKernel);
 }
@@ -142,69 +144,10 @@ __global__ void ROIPoolBackwardAccKernel(const int count, const Dtype* top_diff,
   for (int index = (blockIdx.x + blockIdx.y * gridDim.x) * blockDim.x + threadIdx.x;
        index < count;
        index += blockDim.x * gridDim.x * gridDim.y) {
-    // (n, c, h, w) coords in bottom data
-    int w = index % width;
-    int h = (index / width) % height;
-    int c = (index / width / height) % channels;
-    int n = index / width / height / channels;
-
-    Dtype gradient = 0;
-    // Accumulate gradient over all ROIs that pooled this element
-    for (int roi_n = 0; roi_n < num_rois; ++roi_n) {
-      const Dtype* offset_bottom_rois = bottom_rois + roi_n * 5;
-      int roi_batch_ind = offset_bottom_rois[0];
-      // Skip if ROI's batch index doesn't match n
-      if (n != roi_batch_ind) {
-        continue;
-      }
-
-      int roi_start_w = round(offset_bottom_rois[1] * spatial_scale);
-      int roi_start_h = round(offset_bottom_rois[2] * spatial_scale);
-      int roi_end_w = round(offset_bottom_rois[3] * spatial_scale);
-      int roi_end_h = round(offset_bottom_rois[4] * spatial_scale);
-
-      // Skip if ROI doesn't include (h, w)
-      const bool in_roi = (w >= roi_start_w && w <= roi_end_w &&
-                           h >= roi_start_h && h <= roi_end_h);
-      if (!in_roi) {
-        continue;
-      }
-
-      int offset = (roi_n * channels + c) * pooled_height * pooled_width;
-      const Dtype* offset_top_diff = top_diff + offset;
-      const Dtype* offset_argmax_data = argmax_data + offset;
-
-      // Compute feasible set of pooled units that could have pooled
-      // this bottom unit
-
-      // Force malformed ROIs to be 1x1
-      int roi_width = max(roi_end_w - roi_start_w + 1, 1);
-      int roi_height = max(roi_end_h - roi_start_h + 1, 1);
-
-      Dtype bin_size_h = static_cast<Dtype>(roi_height)
-                         / static_cast<Dtype>(pooled_height);
-      Dtype bin_size_w = static_cast<Dtype>(roi_width)
-                         / static_cast<Dtype>(pooled_width);
-
-      int phstart = floor(static_cast<Dtype>(h - roi_start_h) / bin_size_h);
-      int phend = ceil(static_cast<Dtype>(h - roi_start_h + 1) / bin_size_h);
-      int pwstart = floor(static_cast<Dtype>(w - roi_start_w) / bin_size_w);
-      int pwend = ceil(static_cast<Dtype>(w - roi_start_w + 1) / bin_size_w);
-
-      phstart = min(max(phstart, 0), pooled_height);
-      phend = min(max(phend, 0), pooled_height);
-      pwstart = min(max(pwstart, 0), pooled_width);
-      pwend = min(max(pwend, 0), pooled_width);
-
-      for (int ph = phstart; ph < phend; ++ph) {
-        for (int pw = pwstart; pw < pwend; ++pw) {
-          if (static_cast<int>(offset_argmax_data[ph * pooled_width + pw]) == (h * width + w)) {
-            gradient += offset_top_diff[ph * pooled_width + pw];
-          }
-        }
-      }
+    int max_idx = static_cast<int>(argmax_data[index]);
+    if (max_idx >= 0) {
+      atomicAdd(&bottom_diff[max_idx], top_diff[index]);
     }
-    bottom_diff[index] += gradient;
   }
 }
 
@@ -218,7 +161,7 @@ inline void ROIPoolBackwardAcc(const Tensor<gpu, 4, Dtype> &in_grad,
   const Dtype *bottom_rois = bbox.dptr_;
   Dtype *bottom_diff = in_grad.dptr_;
   Dtype *argmax_data = max_idx.dptr_;
-  const int count = in_grad.shape_.Size();
+  const int count = out_grad.shape_.Size();
   const int num_rois = bbox.size(0);
   const int channels = in_grad.size(1);
   const int height = in_grad.size(2);