@sxjscience 2019-07-12
I think it should be good. One concern is it may be backward incompatible due to the `atomicAdd`. Although I feel it's reasonable to use the faster version, I need to confirm with the GluonCV team whether it will break the current training scripts.