@zhreshold 2019-07-12
Would be nice to apply to ROIAlign similarly since GluonCV have transitioned to use ROIAlign in recent networks.