@larroy 2019-07-12
This is great, is the performance increase only due to type changes as in the description?