@wkcn 2019-07-13
@larroy I think the main reason is that the indices of the maximum numbers are saved in this PR.