@KellenSunderland 2019-07-16
LGTM.  Many thanks for the speedup / refactor.